requirements.txt
pandas==1.0.4
mysql-connector-python==8.0.20
psycopg2-binary==2.8.5
loguru==0.5.1
